Replaced all Long ids with Int ids (as SQLite requires)

+++ app/src/main/java/ru/nsu/fit/data/model/Bicycle.kt
@@ -16,8 +16,8 @@ data class Bicycle(
     val picture: ByteArray,
 
     //references
-    val typeId: Long,
-    val stateId: Long,
-    val wheelSizeId: Long,
-    val colorId: Long,
+    val typeId: Int,
+    val stateId: Int,
+    val wheelSizeId: Int,
+    val colorId: Int,
 )
