+++ app/src/main/java/ru/nsu/fit/data/model/BicycleIssueXref.kt
@@ -4,6 +4,6 @@ import androidx.room.Entity
 
 @Entity(primaryKeys = ["bikeIdRef", "issueIdRef"])
 data class BicycleIssueXref(
-    val bikeIdRef: Long,
-    val issueIdRef: Long
+    val bikeIdRef: Int,
+    val issueIdRef: Int
 )
