+++ app/src/main/java/ru/nsu/fit/data/model/BicycleState.kt
@@ -5,6 +5,6 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName="bicycle_states")
 data class BicycleState(
-    @PrimaryKey(autoGenerate = true) val stateId: Long,
+    @PrimaryKey(autoGenerate = true) val stateId: Int,
     val stateName: String
 )
