+++ app/src/main/java/ru/nsu/fit/data/model/Sale.kt
@@ -5,7 +5,7 @@ import androidx.room.PrimaryKey
 
 @Entity
 data class Sale(
-    @PrimaryKey(autoGenerate = true) val bicycleId: Long,
+    @PrimaryKey(autoGenerate = true) val bicycleId: Int,
     val saleDate: String,
     val customer: Long,
     val finalCost: Long
